requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
schedule>=1.2.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.0.0
